fix error and add reqirement table and mapper

+++ src/Kalabean.Domain/Responses/OrderDetailResponse.cs
@@ -10,10 +10,10 @@ namespace Kalabean.Domain.Responses
     {
         public long Id { get; set; }
         public long OrderId { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public int Num { get; set; }
         public int Price { get; set; }
-        public ThumbResponse<int> ProductThumb { get; set; }
+        public ThumbResponse<long> ProductThumb { get; set; }
         public OrderHeaderResponse OrderHeader { get; set; }
     }
 }
